requirements.txt
requests>=2.25.0
urllib3>=1.26.0
elasticsearch>=8.0.0
python-dotenv==1.2.1
